+++ src/pages/members_page.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import UserCard from './user_card';
+import UserCard from '../components/cards/user_card';
 import { Grid, Row, Col } from 'react-flexbox-grid';
 import axios from 'axios';
-import env from './environment';
+import env from '../components/environment';
 
 
 class Members extends React.Component {
